+++ channel_stats/README.md
@@ -1,9 +1,9 @@
-# RapidPro Apps
+# RapidPro Apps: Channel Stats
 
-To work properly, the `channel_stats` app must be in `PROJECT_ROOT/apps` directory. Once in it, include on project 
+To work properly, the `channel_stats` app must be in `$PROJECT_ROOT/apps` directory. Once in it, include on project 
 settings:
 
-Settings:
+**settings.py**
 ```python
 from .settings_common import *
 
@@ -15,14 +15,3 @@ INSTALLED_APPS = INSTALLED_APPS + (
 ```
 
 This will allow to automatically include the urls in API scope: `temba.api.v2.urls.urlpatterns`.
-
-The final step is migrate
-
-```shell
-$ python manage.py migrate channel_stats
-```
-
-> **Note:** It's possible to add this app as a sub module using the git.
-> `$ git submodule add <remote-address> apps`
----
-
